Refactored grains code into its own submodule with separate folders
