+++ setup.py
@@ -45,7 +45,8 @@ package_names = [
     'mediagrains.tools',
     'mediagrains.patterngenerators',
     'mediagrains.patterngenerators.video',
-    'mediagrains.patterngenerators.audio'
+    'mediagrains.patterngenerators.audio',
+    'mediagrains.grains'
 ]
 packages = {
     pkg: pkg.replace('.', '/') for pkg in package_names
